[feat/#82]: 메인 컴포넌트 분리
