+++ components/ServiceMain.module.css
@@ -1,0 +1,56 @@
+.main_visual {
+  position: relative;
+  height: 175px;
+  background-image: radial-gradient(239.77% 239.77% at 73.8% 38.92%, #f5f5f5 0%, #acf914 100%);
+  border-radius: 0 30px 0 0;
+}
+.main_visual::before {
+  content: '';
+  display: block;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  margin-left: -215px;
+  margin-top: -97.5px;
+  width: 430px;
+  height: 195px;
+  background-image: url(/imgs/img_main_text.png);
+}
+.main_visual .title {
+  position: relative;
+  padding: 28px;
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
+  z-index: 5;
+}
+.main_visual .title h1 {
+  font-size: 64px;
+  font-weight: 900;
+}
+.main_visual .hashtag ul li {
+  position: absolute;
+  display: inline-block;
+  padding: 0 15px;
+  height: 24px;
+  line-height: 24px;
+  font-size: 12px;
+  font-weight: 700;
+  background-color: var(--point-lime-3);
+  border-radius: 12px;
+  z-index: 3;
+}
+.main_visual .hashtag ul li.tag_1 {
+  left: 80px;
+  bottom: 65px;
+}
+.main_visual .hashtag ul li.tag_2 {
+  left: 0;
+  bottom: 35px;
+  border-radius: 0 12px 12px 0;
+}
+.main_visual .hashtag ul li.tag_3 {
+  left: 30px;
+  bottom: 0;
+  border-radius: 12px 12px 0 0;
+}
